move response inside routes
