+++ src/routes/response/accepted_upload.rs
