+++ src/routes/response/blob_deleted.rs
