+++ src/routes/response/content_info.rs
@@ -3,8 +3,8 @@ use axum::http::request::Parts;
 use axum::http::StatusCode;
 use tracing::{event, Level};
 
-use crate::registry::blob_storage::ContentInfo;
-use crate::response::errors::Error;
+use super::errors::Error;
+pub use crate::registry::blob_storage::ContentInfo;
 
 /**
  * ContentInfo should always be wrapped an Option in routes to avoid failure returns.
