+++ src/routes/response/manifest_deleted.rs
