+++ src/routes/response/manifest_reader.rs
