+++ src/routes/response/verified_manifest.rs
